Fixed the light computation with Fixed32
Doing a cross product then a normalisation lead to an averflow in the magnitude calculation because we squared twice a value, resulting in an overflow or underflow. Fixed by normalising the vectors before calculating the cross product. Post cross normalisation still required because of rounding erros, even with floats.

+++ engine/lib/linal/vector/Vector3.cpp
@@ -95,7 +95,7 @@ namespace ue
    */
   Real Vector3::norm()
   {
-    return sqrtf(x*x + y*y + z*z);
+    return sqrtf((float)(x*x + y*y + z*z));
   }
 
   /*
